requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

matplotlib==3.9.0
mlxtend==0.23.1
numpy==1.26.4
pillow==10.3.0
scikit-learn==1.5.0
timm==1.0.3
# torch==2.2.2+cu121
torchinfo==1.8.0
torchmetrics==1.4.0.post0
# torchvision==0.17.2+cu121
tqdm==4.66.4

# torch==2.2.2 torchvision==0.17.2 --index-url https://download.pytorch.org/whl/cu121